#1044a: split pkg/dataplane/compiler.go (3,509 → 1,663 LOC)
compiler.go held 46 functions across 3,509 LOC. Two function clusters
dominated: NAT family (compileNAT, compileStaticNAT, compileNPTv6,
compileNAT64, nptv6Adjustment, resolveSNATMatchAddr) and the firewall-
filter family (compileFirewallFilters, expandFilterTerm, setFilterAddr,
computeFilterProtoPrefilter, resolvePortRange, forwardingClassToDSCP,
resolvePortName).

Extracted both clusters into per-domain files, including each
function's preceding doc comment so go doc still associates them
correctly:
- compiler_nat.go: 1,111 LOC across 6 fns
- compiler_filter.go: 722 LOC across 7 fns

After split:
- compiler.go: 3,509 → 1,663 LOC (53% reduction)
- 2 new files, both well under the 2,000 LOC modularity threshold

`goimports -w` to prune per-file imports. `go build ./...` clean.
`go test ./pkg/dataplane/...` pass.
